requirements.txt
yfinance==0.2.38
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
requests==2.31.0
python-dotenv==1.0.1
edgar==5.4.0
ta==0.11.0
streamlit==1.35.0
plotly==5.22.0
tqdm==4.66.4
joblib==1.4.2
